Move RaptureHotbarModule things around
- Add new `RaptureHotbarModule.ClearCallback` struct and vfuncs
- Move sub-structs out of `RaptureHotbarModule.cs`, separating by concern area.
- Properly express `Client::Game::UI::Hotbar`'s inheritance
