+++ FFXIVClientStructs/FFXIV/Client/Game/UI/Hotbar.cs
@@ -5,29 +5,18 @@ namespace FFXIVClientStructs.FFXIV.Client.Game.UI;
 [StructLayout(LayoutKind.Explicit, Size = 0x08)]
 [VTableAddress("48 8D 05 ?? ?? ?? ?? 48 8B D9 48 89 01 48 81 C1 ?? ?? ?? ?? E8 ?? ?? ?? ?? 48 8D 05 ?? ?? ?? ?? C6 83", 3)]
 public unsafe partial struct Hotbar {
-    [FieldOffset(0x0)] public void* vtbl;
+    [FieldOffset(0x0), Obsolete("Use ClearCallback->vtbl")] public void* vtbl;
+    [FieldOffset(0x0)] public RaptureHotbarModule.ClearCallback* ClearCallback;
 
-    /// <summary>
-    /// Resets <b>ALL</b> hotbars for the current player to their default states. <b>Will reset the hotbars written to disk!</b>
-    /// </summary>
-    /// <param name="raptureHotbarModule">An instance of RaptureHotbarModule.</param>
-    /// <returns>Returns true</returns>
+    /// <inheritdoc cref="RaptureHotbarModule.ClearCallback.ResetAllHotbars"/>
     [VirtualFunction(1)]
     public partial bool ResetAllHotbars(RaptureHotbarModule* raptureHotbarModule);
 
-    /// <summary>
-    /// Resets the PvE hotbars for the specified class at that class' current level. <b>Will reset the hotbars written to disk!</b>
-    /// </summary>
-    /// <param name="raptureHotbarModule">An instance of RaptureHotbarModule.</param>
-    /// <param name="classJobId">The ClassJobID to reset.</param>
-    /// <returns>Returns true.</returns>
+    /// <inheritdoc cref="RaptureHotbarModule.ClearCallback.ResetPvEHotbarsForClass"/>
     [VirtualFunction(2)]
     public partial bool ResetPvEHotbarsForClass(RaptureHotbarModule* raptureHotbarModule, int classJobId);
 
-    /// <inheritdoc cref="ResetPvPHotbarsForClassInner"/>
-    /// <param name="classJobId">The ClassJobId to reset.</param>
-    /// <param name="raptureHotbarModule">An instance of RaptureHotbarModule.</param>
-    /// <returns>Returns true.</returns>
+    /// <inheritdoc cref="RaptureHotbarModule.ClearCallback.ResetPvPHotbarsForClass"/>
     [VirtualFunction(3)]
     public partial byte ResetPvPHotbarsForClass(RaptureHotbarModule* raptureHotbarModule, uint classJobId);
